Fixed timestamp handling

+++ src/uk/ac/dundee/computing/aec/lib/Convertors.java
@@ -1,6 +1,7 @@
 package uk.ac.dundee.computing.aec.lib;
 
 import java.net.URLDecoder;
+import java.util.Calendar;
 import java.util.Date;
 import java.util.StringTokenizer;
 import java.util.UUID;
@@ -109,12 +110,13 @@ public static String[] SplitTags(String Tags){
 
   }
   
-  static final int EPOCH_DIFFERENCE = 141427 /*days*/ * 24 /*hours*/ * 60 /*minutes*/ * 60 /* seconds */ * 1000;
+static final long EPOCH_DIFFERENCE = (long) 141427 /*days*/ * 24 /*hours*/ * 60 /*minutes*/ * 60 /* seconds */ * 1000;
   
-  public static Date UUIDToDate(UUID u)
+ 
+public static Date UUIDToDate(UUID u)
   {
 	  long origTime = u.timestamp();
-	  long inMilliseconds = origTime / 10000;
+	  long inMilliseconds = origTime / 10000; 
 	  long since1970 = inMilliseconds - EPOCH_DIFFERENCE;
 	  
 	  return new Date(since1970);
